Time: 50 ms (41.05%) | Memory: 73.8 MB (20.78%) - LeetSync

+++ 435-non-overlapping-intervals/non-overlapping-intervals.java
@@ -1,7 +1,7 @@
 class Solution {
     public int eraseOverlapIntervals(int[][] nums) {
         int n = nums.length;
-        Arrays.sort(nums, Comparator.comparingDouble(o -> o[1]));
+        Arrays.sort(nums, Comparator.comparingInt(o -> o[1]));
         int last = nums[0][1];
         int count = 1;
         for(int i = 1; i < n; i++){
